Two major changes:
First:
The Wasm code is now in the main Prvt app, but only partly implemented: only requests for /file are supported for now.
In-browser E2EE with Wasm can be enabled on-demand, for now via a global function.

Second:
Added a Service Worker to the main app. The SW is necessary for when Wasm is used, but in all cases it will cache static assets.